requirements.txt
certifi==2020.6.20
chardet==3.0.4
idna==2.7
joblib==0.14.1
numpy==1.15.4
PyYAML==5.3.1
requests==2.20.0
urllib3==1.24.3
